refactor(validation): rename ValidationError files to kebab-case convention (#21)
